convert standalone-build-test to python
